+++ src/main/java/org/ow2/authzforce/core/pdp/impl/combining/PermitOverridesAlg.java
@@ -17,6 +17,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import oasis.names.tc.xacml._3_0.core.schema.wd_17.DecisionType;
+
 import org.ow2.authzforce.core.pdp.api.BaseCombiningAlg;
 import org.ow2.authzforce.core.pdp.api.CombiningAlg;
 import org.ow2.authzforce.core.pdp.api.CombiningAlgParameter;
@@ -26,9 +28,9 @@ import org.ow2.authzforce.core.pdp.api.EvaluationContext;
 import org.ow2.authzforce.core.pdp.impl.BaseDecisionResult;
 
 /**
- * This is the standard Permit-Overrides policy/rule combining algorithm. It allows a single evaluation of Permit to take precedence over any number of deny,
- * not applicable or indeterminate results. Note that since this implementation does an ordered evaluation, this class also supports the
- * Ordered-Permit-Overrides algorithm.
+ * This is the standard Permit-Overrides policy/rule combining algorithm. It allows a single evaluation of Permit to
+ * take precedence over any number of deny, not applicable or indeterminate results. Note that since this implementation
+ * does an ordered evaluation, this class also supports the Ordered-Permit-Overrides algorithm.
  */
 public final class PermitOverridesAlg extends BaseCombiningAlg<Decidable>
 {
@@ -46,12 +48,22 @@ public final class PermitOverridesAlg extends BaseCombiningAlg<Decidable>
 		public DecisionResult eval(EvaluationContext context)
 		{
 			/*
-			 * Replaces and enhances atLeastOneError from XACML spec. atLeastOneError == true <=> firstIndeterminateResult != null
+			 * Replaces atLeastOneErrorDP from XACML spec. atLeastOneErrorDP == true <=> firstIndeterminateDPResult !=
+			 * null
 			 */
-			DecisionResult firstIndeterminateResult = null;
+			DecisionResult firstIndeterminateDPResult = null;
+			/*
+			 * Replaces atLeastOneErrorD from XACML spec. atLeastOneErrorD == true <=> firstIndeterminateDResult != null
+			 */
+			DecisionResult firstIndeterminateDResult = null;
+			/*
+			 * Replaces atLeastOneErrorP from XACML spec. atLeastOneErrorP == true <=> firstIndeterminatePResult != null
+			 */
+			DecisionResult firstIndeterminatePResult = null;
 
 			/*
-			 * Replaces and enhances atLeastOneDeny from XACML spec. atLeastOneDeny == true <=> combinedDenyResult != null
+			 * Replaces and enhances atLeastOneDeny from XACML spec. atLeastOneDeny == true <=> combinedDenyResult !=
+			 * null
 			 */
 			DecisionResult combinedDenyResult = null;
 
@@ -75,9 +87,32 @@ public final class PermitOverridesAlg extends BaseCombiningAlg<Decidable>
 					break;
 				case INDETERMINATE:
 					/*
-					 * FIXME: implement extended Indeterminate decisions (result differs if Indeterminate{P} or Indeterminate{D})
+					 * Save Extended Indeterminate value if this is a new type of such value, till the end because
+					 * needed to compute final Extended Indeterminate value
 					 */
-					firstIndeterminateResult = result;
+					switch (result.getExtendedIndeterminate())
+					{
+					case INDETERMINATE:
+						if (firstIndeterminateDPResult == null)
+						{
+							firstIndeterminateDPResult = result;
+						}
+						break;
+					case DENY:
+						if (firstIndeterminateDResult == null)
+						{
+							firstIndeterminateDResult = result;
+						}
+						break;
+					case PERMIT:
+						if (firstIndeterminatePResult == null)
+						{
+							firstIndeterminatePResult = result;
+						}
+						break;
+					default:
+
+					}
 					break;
 				default:
 					break;
@@ -85,11 +120,22 @@ public final class PermitOverridesAlg extends BaseCombiningAlg<Decidable>
 			}
 
 			/*
-			 * FIXME: implement extended Indeterminate decisions as the algorithm distinguishes them.
+			 * There was no Permit, else: if any Indeterminate{DP}, then Indeterminate{DP}
 			 */
-			if (firstIndeterminateResult != null)
+			if (firstIndeterminateDPResult != null)
 			{
-				return firstIndeterminateResult;
+				// at least one Indeterminate{DP}
+				return firstIndeterminateDPResult;
+			}
+
+			/*
+			 * Else if any Indeterminate{P}, then: if ( any Indeterminate{D} or any Deny ) -> Indeterminate{DP}; else ->
+			 * Indeterminate{P} (this is a simplified equivalent of the algo in the spec)
+			 */
+			if (firstIndeterminatePResult != null)
+			{
+				return new BaseDecisionResult(firstIndeterminatePResult.getStatus(), firstIndeterminateDResult != null
+						|| combinedDenyResult != null ? DecisionType.INDETERMINATE : DecisionType.PERMIT);
 			}
 
 			/*
@@ -100,13 +146,19 @@ public final class PermitOverridesAlg extends BaseCombiningAlg<Decidable>
 				return combinedDenyResult;
 			}
 
+			if (firstIndeterminateDResult != null)
+			{
+				return firstIndeterminateDResult;
+			}
+
 			return BaseDecisionResult.NOT_APPLICABLE;
 		}
 
 	}
 
 	@Override
-	public Evaluator getInstance(List<CombiningAlgParameter<? extends Decidable>> params, List<? extends Decidable> combinedElements)
+	public Evaluator getInstance(List<CombiningAlgParameter<? extends Decidable>> params,
+			List<? extends Decidable> combinedElements)
 	{
 		return new Evaluator(combinedElements);
 	}
@@ -114,7 +166,8 @@ public final class PermitOverridesAlg extends BaseCombiningAlg<Decidable>
 	/**
 	 * The standard URN used to identify this algorithm
 	 */
-	static final String[] SUPPORTED_IDENTIFIERS = { "urn:oasis:names:tc:xacml:3.0:policy-combining-algorithm:permit-overrides",
+	static final String[] SUPPORTED_IDENTIFIERS = {
+			"urn:oasis:names:tc:xacml:3.0:policy-combining-algorithm:permit-overrides",
 			"urn:oasis:names:tc:xacml:3.0:rule-combining-algorithm:permit-overrides",
 			"urn:oasis:names:tc:xacml:3.0:policy-combining-algorithm:ordered-permit-overrides",
 			"urn:oasis:names:tc:xacml:3.0:rule-combining-algorithm:ordered-permit-overrides" };